FIXING carrinho and itemcarrinho crud methods issues

+++ src/main/java/app/service/CarrinhoService.java
@@ -1,6 +1,7 @@
 package app.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -8,6 +9,8 @@ import org.springframework.stereotype.Service;
 import app.entity.Carrinho;
 import app.entity.ItemCarrinho;
 import app.repository.CarrinhoRepository;
+import app.repository.ItemCarrinhoRepository;
+import app.repository.ProdutoRepository;
 import jakarta.validation.Valid;
 
 @Service
@@ -15,7 +18,14 @@ public class CarrinhoService {
 
 	@Autowired
 	private CarrinhoRepository carrinhoRepository;
+	@Autowired
+	private ItemCarrinhoRepository itemCarrinhoRepository;
+	@Autowired
+	private ItemCarrinhoService itemCarrinhoService;
+	@Autowired
+	private ProdutoRepository produtoRepository;
 	
+	//metodo para calcular o valor final do carrinho automaticamente 
 	public double valorTotalCarrinho(@Valid List<ItemCarrinho> itemCarrinho) {
 		
 		double valorTotal = 0;
@@ -29,8 +39,24 @@ public class CarrinhoService {
 	}
 	
 	public String save (Carrinho carrinho) {
-		this.carrinhoRepository.save(carrinho);
-		return carrinho.getValorCarrinho() + "  registrada";
+		// Percorre os itens do carrinho e os salva no banco de dados
+        for (ItemCarrinho item : carrinho.getItemCarrinho()) {
+            item.setCarrinho(carrinho);
+          //adiciona o valor do produto a variavel valorUnitario para manter o registro do valor na compra
+    		double valorUnitario = this.itemCarrinhoService.setValor(item.getProduto());
+    		item.setValorUnitario(valorUnitario);
+            // Salva o item do carrinho no banco de dados
+            itemCarrinhoRepository.save(item);
+            // Salva o produto associado ao item do carrinho no banco de dados
+            produtoRepository.save(item.getProduto());
+        }
+        //chamada do metodo para fazer o calculo do valor final do carrinho antes de persistir o mesmo
+        double valorFinal = this.valorTotalCarrinho(carrinho.getItemCarrinho());
+        carrinho.setValorCarrinho(valorFinal);
+        //salva o carrinho no banco de dados 
+        this.carrinhoRepository.save(carrinho);
+    
+        return carrinho.getValorCarrinho() + "  registrada";
 	}
 	
 	public List <Carrinho> listAll () {
@@ -44,12 +70,53 @@ public class CarrinhoService {
 	}
 	
 	public String update (Carrinho carrinho, long idCarrinho) {
+		// Define o ID do carrinho com base no ID fornecido
 		carrinho.setIdCarrinho(idCarrinho);
+		
+		//verifica se o id  mandado pelo usuario e valido, caso nao seja devolve uma exception
+		Optional<Carrinho> car = this.carrinhoRepository.findById(idCarrinho);
+		if (car.isEmpty()) {
+			throw new RuntimeException("Id nao encontrado");
+		}
+	
+		// Percorre os itens do carrinho e os salva no banco de dados
+		for (ItemCarrinho item : carrinho.getItemCarrinho()) {
+            item.setCarrinho(carrinho);
+            
+         // Verifica se o item já existe no banco de dados
+            if (item.getIdItem() != 0 && itemCarrinhoRepository.existsById(item.getIdItem())) {
+                // Se existe, atualiza o item
+                itemCarrinhoRepository.save(item);
+            } else {
+                // Se não existe, salva como um novo item
+                itemCarrinhoRepository.save(item);
+            }
+            //adiciona o valor do produto a variavel valorUnitario para manter o registro do valor na compra
+    		double valorUnitario = this.itemCarrinhoService.setValor(item.getProduto());
+    		item.setValorUnitario(valorUnitario);
+            
+        }
+		//chamada do metodo para fazer o calculo do valor final do carrinho antes de persistir o mesmo
+		double valorFinal = this.valorTotalCarrinho(carrinho.getItemCarrinho());
+		carrinho.setValorCarrinho(valorFinal);	
 		this.carrinhoRepository.save(carrinho);
-		return "O " + carrinho.getValorCarrinho() + " Venda atualizada";
+		return " Carrinho " + carrinho.getValorCarrinho() + " Foi atualizado";
+		
 	}
 	
 	public String delete (long idCarrinho) {
+		//verifica se o id  mandado pelo usuario e valido, caso nao seja devolve uma exception
+		Optional<Carrinho> car = this.carrinhoRepository.findById(idCarrinho);
+		if (car.isEmpty()) {
+			throw new RuntimeException("Id nao encontrado");
+		}
+		Carrinho carrinho = car.get();
+		// Percorre os itens do carrinho e os salva no banco de dados
+        for (ItemCarrinho item : carrinho.getItemCarrinho()) {
+            item.setCarrinho(carrinho);
+            // Deleta o item do carrinho no banco de dados
+            itemCarrinhoRepository.deleteById(item.getIdItem());
+        }
 		this.carrinhoRepository.deleteById(idCarrinho);
 		return " Venda deletada";
 	}
